refactor : Board 생성 위치를 ChessGame으로 변경
