+++ src/main/java/chess/controller/ChessController.java
@@ -1,5 +1,6 @@
 package chess.controller;
 
+import chess.domain.ChessGame;
 import chess.domain.board.Board;
 import chess.view.InputView;
 import chess.view.OutputView;
@@ -17,8 +18,8 @@ public class ChessController {
             if (commend.equals("end")) {
                 break;
             }
-            Board board = new Board();
-            OutputView.printChessBoard(board);
+            ChessGame chessGame = new ChessGame();
+            OutputView.printChessBoard(chessGame.getBoard());
         }
     }
 }
